Add files via upload
This Lab involves 20 Nodes using CISCO CML2.  I configured R1-R3 & R8 with protocols OSPF and BGP configuration. I then configured R4 - R7 with protocols EIGRP and BGP configuration. Finally, I configurated R9 - R11 with OSPF, and R11 has a default gateway to access the internet. Don't forget to Configure the default gateway to the PCs and assign the IP address of eth0.  This was a fun project and helped with troubleshooting.